requirements.txt
tk
googletrans==4.0.0-rc1
gTTS
playsound
SpeechRecognition
pillow